Fix task interface (#64)
This PR includes several related changes/fixes:
* Removed previous Task class and changed TaskInstance to Task.
* Removed the needs for `clear_created_task_types()` and
`register_task_type()` methods.
* Removed the support for task filtering by type when instantiating the
Benchmark. The reason we needed it in the first place was because the
Extraction task was slow to instantiate (because it imports the
sentence_transformers package). Ideally, no task should be costly to
load and each Task implementation should be done carefully to perform
any costly initialization in a lazy fashion (i.e., just-in-time when it
is required). Correspondingly, I have updated the extraction Task to
lazily import `kebab.tasks.extraction.metrics.aesop.calculator` which in
turn imports the sentence_transformers package. However, I was unable
reproduce the slow downs Pavel and I observed on his machine. So, we
might need to re-test this on his machine after he is back from leave.
* Moved instances.json and REBEL schema from DiSK repo to MS-KeBAB repo.

The following PR contains related changes for the DiSK repo that will be
merged after this PR is merged:
https://github.com/microsoft/DiSK/pull/31